Change amount measured to decimal

+++ Rainfall.Application/Models/RainfallReading.cs
@@ -6,6 +6,6 @@
     public class RainfallReading
     {
         public DateTime DateMeasured { get; set; }
-        public double AmountMeasured { get; set; }
+        public decimal AmountMeasured { get; set; }
     }
 }
